feat: add climbs to topo
Adding climbs to a topo effectively means that climb can be described by
that topo in some capacity. This commit introduces the idea of a session
context. These climbs begin as a part of the session context (described
later) and can be added to the world context.

The session context is introduced as an ephemeral state, only present
for the duration of the editor. In this commit, the session state is
populated with all climbs by the new-topo page, but later that can be
changed to something a bit more granular.