feat: ability to create spline geometries from climb
Adds a tool property to the session context. Creates a "create spline
tool" that creates splines when it receives mouse events. This tool
produces a spline through a pub-sub pattern. Creates a tool per-climb in
the properties panel. The panel 1. subscribes to the tool to edit the
world context when the tool emits a spline; and 2. loads the tool into
the session context when a button is clicked.

This tool is defined _exactly_ how it is needed with no regard to
extensibility. Refactors will need to occur to reduce repeated code as
more tools are added.