refactor!: make topos line-centric
Makes topos more line-centric (and less climb-centric). Topos can have
the idea of "lines," not climbs. In a future patch, a property can be
added to lines that can add a climb-id attribute to a line.

This has the consequence of changing how the properties panel is
organized, making it less nested. During this "rewrite" of the panel, I
discovered a bit more about client-components. I was able to remove most
of the "use-client" directives as they are all loaded under a "client
tree." This allows me to pass non-serializable things down. I _think_ if
I try to use this on a boundary, I will get an error, in which can I
think I can add both `"use client"` and `import "client-only"` to
enforce how I think these components should behave. But for now, they
are not needed, so I will keep them as-is.