refactor: use a use-image hook
This pulls out the image-loading logic into a hook. This hook takes a
source string and loads it into an HTML image element. Once the image
loads, the 0-index will contain the HTML image element, ready to be
used. This patch re-implements the image logic in `CanvasArea` and
`BackgroundProperties` with this new hook.

Notably, this patch removes the sloppy image dimension initialization
and replaces it with a button to fit the image explicitly. The image is
initially has no size, which can be somewhat confusing.