feat: create topo by uploading image
To bring to a simpler feature-pairty with my current website, this patch
adds to page with a form to upload an image "as a topo." Upon
submission, the user will be redirected to the topo editor to edit the
newly created topo. Behind the scenes, this uploads and image, creates a
topo (with the image dimensions), and adds that image to the topo.