todo! fix: size paths to normalized world
Right now the paths are sized to the world.. this is a problem because
the world could be sized to the background image.. if that image has a
much larger size then the world will be much larger and the paths will
appear much smaller. If I have a 4000x3000 image and a 400x300 canvas,
the paths should look the same between them. Allowing modification of
the path widths of the topos lines seems like too much control that I do
not wish for the user to control. It'd make it harder to ensure
stylistic consistency between topos.