todo! a correction to a couple previous todos
If I always tell the user to fit to the 4000x3000 canvas.. that is it
will either be 4000 max width or 3000 max height, then I can continue
using the coordinates/sizing I am now, but that comes with the
consequence of have potentially weird world sizes, because that would
then be sized to the image and likely have a fractional size.

I'm not sure what is better at this point. I'm almost leaning toward
using the image's natural size as the world size.. but then I am back
to the "topo must have an image" rule that I was hoping to avoid.